+++ src/axom/core/execution/internal/cuda_exec.hpp
@@ -65,6 +65,18 @@ struct execution_space<CUDA_EXEC<BLOCK_SIZE, SYNCHRONOUS>>
   {
     return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
   }
+  static constexpr runtime_policy::Policy runtimePolicy() noexcept
+  {
+    return runtime_policy::Policy::cuda;
+  }
+  static bool usesMemorySpace(axom::MemorySpace m) noexcept
+  {
+    return m == memory_space || m == MemorySpace::Unified;
+  }
+  static bool usesAllocId(int allocId) noexcept
+  {
+    return usesMemorySpace(axom::detail::getAllocatorSpace(allocId));
+  }
 };
 
 /*!
@@ -94,6 +106,19 @@ struct execution_space<CUDA_EXEC<BLOCK_SIZE, ASYNC>>
   static int allocatorID() noexcept
   {
     return axom::getUmpireResourceAllocatorID(umpire::resource::Device);
+  }
+  static constexpr runtime_policy::Policy runtimePolicy() noexcept
+  {
+    return runtime_policy::Policy::cuda;
+  }
+  static bool usesMemorySpace(axom::MemorySpace m) noexcept
+  {
+    return m == memory_space || m == MemorySpace::Unified;
+  }
+  static bool usesAllocId(int allocId) noexcept
+  {
+    return allocId == 0 ||
+      usesMemorySpace(axom::detail::getAllocatorSpace(allocId));
   }
 };
 }  // namespace axom